Head_hunter_DB_search_project/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
typing-extensions==4.9.0
